chore: move core to config
